picture_balloons/requirements.txt
numpy~=1.24.1
pygame~=2.1.3.dev8
Pillow~=9.4.0